+++ autofit/non_linear/mcmc/emcee/samples.py
@@ -1,14 +1,11 @@
 import emcee
 import numpy as np
 from typing import List, Optional
-import warnings
 
 from autofit.mapper.prior_model.abstract import AbstractPriorModel
 from autofit.non_linear.samples.mcmc import MCMCSamples
 from autofit.non_linear.mcmc.auto_correlations import AutoCorrelationsSettings, AutoCorrelations
 from autofit.non_linear.samples import Sample
-
-from autofit import exc
 
 class EmceeSamples(MCMCSamples):
 
@@ -75,44 +72,6 @@ class EmceeSamples(MCMCSamples):
             unconverged_sample_size=unconverged_sample_size,
             time=time,
             results_internal=results_internal,
-        )
-
-    def __add__(
-            self,
-            other: "EmceeSamples"
-    ) -> "EmceeSamples":
-        """
-        Samples can be added together, which combines their `sample_list` meaning that inferred parameters are
-        computed via their joint PDF.
-
-        For Emcee samples there are no tools for combining results in their native format, therefore these
-        `results_internal` are set to None and support for visualization is disabled.
-
-        Parameters
-        ----------
-        other
-            Another Samples class
-
-        Returns
-        -------
-        A class that combined the samples of the two Samples objects.
-        """
-
-        self._check_addition(other=other)
-
-        warnings.warn(
-            "Addition of EmceeSamples cannot retain results in native format. "
-            "Visualization of summed samples diabled.",
-            exc.SamplesWarning
-        )
-
-        return EmceeSamples(
-            model=self.model,
-            sample_list=self.sample_list + other.sample_list,
-            auto_correlation_settings=self.auto_correlation_settings,
-            unconverged_sample_size=self.unconverged_sample_size,
-            time=self.time,
-            results_internal=None
         )
 
     @property
