+++ autofit/non_linear/nest/ultranest/samples.py
@@ -1,51 +1,12 @@
 from typing import Optional
-import warnings
 
 from autofit.mapper.prior_model.abstract import AbstractPriorModel
 from autofit.non_linear.samples import Sample
 from autofit.non_linear.samples.nest import NestSamples
 
-from autofit import exc
+
 
 class UltraNestSamples(NestSamples):
-
-    def __add__(
-            self,
-            other: "UltraNestSamples"
-    ) -> "UltraNestSamples":
-        """
-        Samples can be added together, which combines their `sample_list` meaning that inferred parameters are
-        computed via their joint PDF.
-
-        For UltraNest samples there are no tools for combining results in their native format, therefore these
-        `results_internal` are set to None and support for visualization is disabled.
-
-        Parameters
-        ----------
-        other
-            Another Samples class
-
-        Returns
-        -------
-        A class that combined the samples of the two Samples objects.
-        """
-
-        self._check_addition(other=other)
-
-        warnings.warn(
-            "Addition of UltraNestSamples cannot retain results in native format. "
-            "Visualization of summed samples diabled.",
-            exc.SamplesWarning
-        )
-
-        return UltraNestSamples(
-            model=self.model,
-            sample_list=self.sample_list + other.sample_list,
-            number_live_points=self._number_live_points,
-            unconverged_sample_size=self.unconverged_sample_size,
-            time=self.time,
-            results_internal=None
-        )
 
     @classmethod
     def from_results_internal(
