+++ autofit/non_linear/samples/nest.py
@@ -1,10 +1,12 @@
 from typing import List, Optional
+import warnings
 
 from autofit.mapper.prior_model.abstract import AbstractPriorModel
 from autofit.non_linear.samples.pdf import PDFSamples
-from .sample import Sample
-from .stored import StoredSamples
+from autofit.non_linear.samples.samples import Sample
+from autofit.non_linear.samples.stored import StoredSamples
 
+from autofit import exc
 
 class NestSamples(PDFSamples):
 
@@ -55,6 +57,44 @@ class NestSamples(PDFSamples):
             unconverged_sample_size=unconverged_sample_size,
             time=time,
             results_internal=results_internal
+        )
+
+    def __add__(
+            self,
+            other: "NestSamples"
+    ) -> "NestSamples":
+        """
+        Samples can be added together, which combines their `sample_list` meaning that inferred parameters are
+        computed via their joint PDF.
+
+        For UltraNest samples there are no tools for combining results in their native format, therefore these
+        `results_internal` are set to None and support for visualization is disabled.
+
+        Parameters
+        ----------
+        other
+            Another Samples class
+
+        Returns
+        -------
+        A class that combined the samples of the two Samples objects.
+        """
+
+        self._check_addition(other=other)
+
+        warnings.warn(
+            f"Addition of {self.__class__.__name__} cannot retain results in native format. "
+            "Visualization of summed samples diabled.",
+            exc.SamplesWarning
+        )
+
+        return self.__class__(
+            model=self.model,
+            sample_list=self.sample_list + other.sample_list,
+            number_live_points=self._number_live_points,
+            unconverged_sample_size=self.unconverged_sample_size,
+            time=self.time,
+            results_internal=None
         )
 
     @property
